Add client credentials and custom profile service to IdentityServer

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -44,12 +45,13 @@ public class AuctionsController(AuctionDbContext context,
         return mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
     {
         var auction = mapper.Map<Auction>(createAuctionDto);
 
-        auction.Seller = "test";
+        auction.Seller = User.Identity?.Name ?? "Unknown user";
 
         context.Auctions.Add(auction);
 
@@ -67,6 +69,7 @@ public class AuctionsController(AuctionDbContext context,
         return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -77,6 +80,11 @@ public class AuctionsController(AuctionDbContext context,
         if (auction is null)
         {
             return NotFound();
+        }
+
+        if (auction.Seller != User.Identity?.Name)
+        {
+            return Forbid();
         }
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
@@ -94,6 +102,7 @@ public class AuctionsController(AuctionDbContext context,
             : BadRequest("Problem saving changes.");
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -104,6 +113,11 @@ public class AuctionsController(AuctionDbContext context,
         if (auction is null)
         {
             return NotFound();
+        }
+
+        if (auction.Seller != User.Identity?.Name)
+        {
+            return Forbid();
         }
 
         context.Auctions.Remove(auction);
